Improve docker container management
- Simplify tagging.
- Test removal of old container.

+++ dmtf/run-redfish-simulator.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
-# Build the docker container first
-docker build -t localhost:5000/redfish-simulator:latest .
-# Th -p option needs to be after the run command. No warning is given if before but doesn't work
-docker rm "redfish-simulator"
-docker run -d -p 8000:80 --name "redfish-simulator" localhost:5000/redfish-simulator:latest
+docker ps -a | grep -q "redfish-simulator:latest"
+if [ "$?" -eq 0 ]; then
+	docker rm "redfish-simulator"
+fi
+# The -p option needs to be after the run command. No warning is given if before but doesn't work
+docker run -d -p 8000:80 --name "redfish-simulator" redfish-simulator:latest
 echo "Launch your browser and load http://localhost:8000/redfish/v1"
